WIP PageNum
* PageNum -> setPosition
* PageNum -> creatFormattedText
* WIP PageNum -> renderAsBitmap

+++ src/main/java/net/sf/mcf2pdf/mcfelements/McfPageNum.java
@@ -18,7 +18,7 @@ public interface McfPageNum {
 	
 	public Color getBgColor();
 	
-	public int getFontSize();
+	public float getFontSize();
 
 	public int getFontBold();
